requirements.txt
ffmpeg-python
Flask
Flask_Cors
gradio>=3.7.0
numpy==1.23.5
pyworld
scipy==1.10.0
SoundFile==0.12.1
torch
torchaudio
torchcrepe
tqdm
rich
loguru
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
fairseq==0.12.2
librosa==0.9.1
tensorboard
tensorboardX
transformers
edge_tts
langdetect
pyyaml
pynvml
faiss-cpu
einops
local_attention
Flask
Flask_Cors
gradio>=3.7.0
#pyworld==0.2.5
scipy==1.10.0
SoundFile==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchcrepe
tqdm
#rich.progress
loguru
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
onnxruntime-gpu
librosa==0.9.1
tensorboard
tensorboardX
edge_tts
langdetect
pyyaml
pynvml
transformers
ffmpeg-python
faiss-cpu
spotipy
youtube_dl
youtube_search
yt_dlp
ffprobe
# install from https://pytorch.org/get-started/locally/
# torch>=1.12.0
# torchvision>=0.6.1
tqdm>=4.30
librosa>=0.9
opencv_python>=4.2.0


